upd ssh

+++ env/$HOME/.jerryc05.env.sh
@@ -30,34 +30,26 @@ for USR in "$HOME" $USR_DIRS; do
 done
 
 
-# SSH and GPG
+# SSH Agent
 DOT_SSH="$HOME/.ssh"
 mkdir -p $DOT_SSH
 
 # If using security key under WSL, with Putty in Windows
-if grep -qEi "(Microsoft|WSL)" /proc/version &>/dev/null && command -v putty.exe >/dev/null 2>&1; then
-  # `(cd ~/.ssh&&curl -JOL https://github.com/BlackReloaded/wsl2-ssh-pageant/releases/latest/download/wsl2-ssh-pageant.exe&&chmod +x ./wsl2-ssh-pageant.exe)`
+PIPERELAY="$DOT_SSH/npiperelay.exe"
+if command -v socat >/dev/null 2>&1 && grep -qEi "(Microsoft|WSL)" /proc/version &>/dev/null && [ -f "$PIPERELAY" ]; then
+  # Download from [https://github.com/jstarks/npiperelay/releases
+  # Copy [npiperelay.exe] to ~/.ssh/
+  # [chmod +x ~/.ssh/npiperelay.exe]
 
-  # SSH Socket
-  # Removing Linux SSH socket and replacing it by link to wsl2-ssh-pageant socket
   export SSH_AUTH_SOCK="$DOT_SSH/agent.sock"
-  ss -a | fgrep -q $SSH_AUTH_SOCK
+  WINDOWS_SSH_PIPE="//./pipe/openssh-ssh-agent"
+  ps -auxww | grep -q "[n]piperelay.exe -ei -s $WINDOWS_SSH_PIPE"
   if [ $? -ne 0 ]; then
-    rm -f $SSH_AUTH_SOCK
-    setsid nohup socat UNIX-LISTEN:$SSH_AUTH_SOCK,fork EXEC:"$DOT_SSH/wsl2-ssh-pageant.exe" &>/dev/null &
-  fi
-
-  # GPG Socket
-  # Removing Linux GPG Agent socket and replacing it by link to wsl2-ssh-pageant GPG socket
-  export GPG_AGENT_SOCK=$HOME/.gnupg/S.gpg-agent
-  ss -a | fgrep -q $GPG_AGENT_SOCK
-  if [ $? -ne 0 ]; then
-    rm -rf $GPG_AGENT_SOCK
-    setsid nohup socat UNIX-LISTEN:$GPG_AGENT_SOCK,fork EXEC:"$DOT_SSH/wsl2-ssh-pageant.exe --gpg S.gpg-agent" &>/dev/null &
+    [ -S $SSH_AUTH_SOCK ] && rm -f $SSH_AUTH_SOCK
+    setsid nohup socat UNIX-LISTEN:$SSH_AUTH_SOCK,fork EXEC:"$PIPERELAY -ei -s $WINDOWS_SSH_PIPE",nofork &>/dev/null &
   fi
 
 else
-  # SSH Agent
   if command -v ssh-agent >/dev/null 2>&1 && command -v ssh-add >/dev/null 2>&1; then
     ssh-add -l >/dev/null 2>&1
     if [ $? -eq 2 ]; then
